backend/requirements.txt
chromadb
openai
dotenv
langchain-community
langchain-text-splitters
pypdf
langchain-openai
tiktoken